refactor(webapp): move account store to runes
