+++ frontend/vitest.config.mjs
@@ -1,11 +1,22 @@
 import path from "node:path";
 import { fileURLToPath } from "node:url";
 
+import { svelte } from "@sveltejs/vite-plugin-svelte";
 import { defineConfig } from "vitest/config";
 
 const __dirname = path.dirname(fileURLToPath(import.meta.url));
+const srcDir = path.resolve(__dirname, "src");
 
 export default defineConfig({
+  plugins: [
+    svelte({
+      dynamicCompileOptions({ filename }) {
+        if (filename && path.resolve(filename).startsWith(srcDir)) {
+          return { runes: true };
+        }
+      },
+    }),
+  ],
   resolve: {
     alias: {
       $lib: path.resolve(__dirname, "src/lib"),
